refactor(cli): fold orchestrate _spec helpers into package __init__
Move _spec.py function bodies directly into orchestrate/__init__.py.
All names remain importable from lionagi.cli.orchestrate. The _argparse
alias is replaced with the already-imported argparse directly.
